movie-consumer/src/main/java/org/example/consumers/MovieConsumer.java
package org.example.consumers;


import com.fasterxml.jackson.core.JsonProcessingException;
import lombok.extern.slf4j.Slf4j;
import org.apache.kafka.clients.consumer.ConsumerRecord;
import org.example.services.MovieService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.kafka.annotation.KafkaListener;
import org.springframework.stereotype.Component;

@Component
@Slf4j
public class MovieConsumer {

    @Autowired
    MovieService movieService;

    @KafkaListener(topics = {"movies"})
    public void onMessage(ConsumerRecord<Integer, String> consumerRecord) throws JsonProcessingException {

        log.info("ConsumerRecord : {}", consumerRecord);
        movieService.processMovie(consumerRecord);
    }
}
